working csvs for both, data partially in ml

+++ run_multiprocessing.py
@@ -1,9 +1,7 @@
 import multiprocessing
-from multiprocessing import Pool
 import workers
 import sys
 import time
-import tkinter as tk
 
 # increase recursion limit
 sys.setrecursionlimit(15000)
@@ -12,6 +10,7 @@ if __name__ == "__main__":
     # turns given process into a daemon which will run forever normally
     # subprocess is automatically terminated after the parent process ends to prevent orphan processes
     # aka kills all subprocesses
+    start_time = time.time()
     proc1 = multiprocessing.Process(target=workers.worker1)
     proc1.daemon = True
     proc2 = multiprocessing.Process(target=workers.worker2)
@@ -20,9 +19,8 @@ if __name__ == "__main__":
     proc1.start() 
     proc2.start()
 
-    while True:
+    if (int(time.time() - start_time)) % 15 == 0.0 and (int(time.time() - start_time)) != 0:
         proc3 = multiprocessing.Process(target=workers.worker3)
-        time.sleep(15)
+        # time.sleep(300) # do process only every 5 min
         proc3.start()
-        proc3.join()
         proc3.terminate()
